backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
opencv-python-headless
pillow
numpy
google-generativeai
